ShareViewController: implement auto link processing First half of share-sheet integration (mostly) complete
